\name{31_new_generics}
\alias{bvplot}
\alias{bvrng}
\alias{bvmat}
\title{New Generic Functions}
\description{New generic functions for plotting function objects, generating random numbers and computing density matrices.}
\usage{
bvplot (\dots)
bvrng (\dots)
bvmat (\dots)
}
\arguments{
\item{\dots}{Arguments for specific methods.} 
}
\details{
Do not call bvplot.\cr
Call plot instead.
}
\references{Refer to the vignette for an overview, references, theoretical background and better examples.}
\seealso{
\link{Main Plotting Functions}

\link{Random Numbers}

\link{Density Matrices}
}
