\name{32_new_generics}
\alias{bvrng}
\alias{bvmat}
\title{New Generic Functions}
\description{New generic functions for generating random numbers and computing density matrices.}
\usage{
bvrng (\dots)
bvmat (\dots)
}
\arguments{
\item{\dots}{Arguments for specific methods.} 
}
\references{Refer to the vignette for an overview, references, theoretical background and better examples.}
\seealso{
\link{Random Numbers}

\link{Density Matrices}
}
