\name{generate}
\alias{generate.bivariate.sample}
\title{Generate Bivariate Samples}
\description{Generate a normal like positively correlated positively skewed sample.}
\usage{
generate.bivariate.sample (n)
}
\arguments{
\item{n}{The sample size.}
}
\examples{
xy = generate.bivariate.sample (200)
}
